createVenueEntity
